requirements.txt
whitenoise==6.9.0
Werkzeug==2.3.7
requests==2.31.0
dnspython==2.4.2
Flask==3.0.2
Flask-Limiter==3.5.0
Pillow==10.2.0